pyproject.toml
[project]
name = "helios-cots-telemetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "betterproto2-compiler>=0.9.0",
    "betterproto2[all]>=0.9.1",
    "cobs>=1.2.2",
    "crcmod>=1.7",
    "helios-python-sdk",
    "protobuf>=7.34.0",
    "pyserial>=3.5",
]

[tool.uv.sources]
helios-python-sdk = { path = "helios-python-sdk", editable = true }
